requirements.txt
click==7.1.2
Cython==0.29.21
Flask==1.1.2
gensim==4.0.1
greenlet==1.0.0
gunicorn==20.1.0
inexactsearch==1.0.2
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
MarkupSafe==1.1.1
nltk==3.6.2
nmslib==2.1.1
numpy==1.20.2
pandas==1.2.4
psutil==5.8.0
pybind11==2.6.1
pyspellchecker==0.6.2
python-dateutil==2.8.1
python-dotenv==0.17.0
pytz==2021.1
regex==2021.4.4
scipy==1.6.2
silpa-common==0.3
six==1.15.0
smart-open==5.0.0
soundex==1.1.3
typing==3.7.4.3
Werkzeug==1.0.1
